- Create a dragAndDrop.js module file - Remove drag&drop from  index.js
